Fuse fmpz_mpoly/is_*
